feat: add ConvertEnumerableToArrayToAllRector
The `toArray()` method on an Enumerable maps over all the items and
recursively calls the `toArray()` method on each item if it is an
Arrayable object. For collections that do not contain any Arrayable
items, this is an unnecessary operation and the method can be replaced
with `all()` which simply returns the underlying array.